Adjust Hikari timeout
Reason this is getting bumped is that I'm seeing a few issues on the Patron bot of timed out connections, but there is no actual connection loss of the ssh tunnel. Let's see if this fixes that.
Bumping this means that in case of an actual connection loss this + the interval of database healthchecks (currently 5 seconds) are the worst case lag a user might see, until the connection is marked as failed and failfasts until it is reestablished. I think having a slight lag at the start of an actual tunnel connection loss (which are rare, about one every 4 days) instead of occasional random timeouts (looks like about 20 per day) is the better compromise.