Add mock credentials to docker image
The reason this is an ok idea is due to the volumes used to mount
the config and credentials files.
If a file is not present when the image is started, docker creates
an empty directory in its place. This is easy to take care outside
of the container, but a bit harder to do so inside of it. Having a
dummy credentials file inside the container makes the process of
fixing it after accidentally starting it without the credentials
files much easier for the selfhosters.